Create base tests
