+++ DigioApp/DigioApp/Modules/ProductList/ProductListConfigurator.swift
