+++ DigioApp/DigioApp/Modules/ProductList/ProductListInteractor.swift
