+++ DigioApp/DigioApp/Modules/ProductList/Views/ProductCell.swift
@@ -7,7 +7,7 @@ final class ProductCell: UICollectionViewCell {
         let imageView = UIImageView()
         imageView.backgroundColor = .clear
         imageView.contentMode = .scaleAspectFit
-        imageView.image = UIImage(named: "noImage")
+        imageView.image = UIImage(named: "empty")
         return imageView
     }()
