--- DigioApp/DigioApp/ProductDetail/ProductDetailRouter.swift
@@ -1,9 +1,0 @@
-import UIKit
-
-final class ProductDetailRouter {
-    weak var viewController: UIViewController?
-}
-
-extension ProductDetailRouter: ProductDetailRouterProtocol {
-    
-}
