+++ DigioApp/DigioApp/Utils/ViewConfiguratorProtocol.swift
