--- DigioApp/DigioApp/ViewController.swift
@@ -1,19 +1,0 @@
-//
-//  ViewController.swift
-//  DigioApp
-//
-//  Created by Ana Carolina Martins Pessoa on 29/01/23.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
-
