+++ DigioApp/DigioAppTests/Modules/ProductDetail/ProductDetailPresenterTests.swift
@@ -1,0 +1,1 @@
+import Foundation
