+++ DigioApp/DigioAppTests/Modules/ProductList/ProductListPresenterTests.swift
@@ -1,0 +1,1 @@
+import Foundation
